cast connection id correctly (solaris complier complains)

+++ client/drizzleslap.cc
@@ -1677,7 +1677,7 @@ static int run_query(drizzle_con_st &con, drizzle_result_st *result,
   if (opt_only_print)
   {
     printf("/* CON: %d */ %.*s;\n",
-           drizzle_context(drizzle_con_drizzle(&con)),
+           (size_t)drizzle_context(drizzle_con_drizzle(&con)),
            len, query);
     return 0;
   }
@@ -1786,7 +1786,7 @@ static void create_schema(drizzle_con_st &con, const char *db, Statement *stmt, 
   if (opt_only_print)
   {
     printf("/* CON: %u */ use %s;\n",
-           drizzle_context(drizzle_con_drizzle(&con)),
+           (size_t)drizzle_context(drizzle_con_drizzle(&con)),
            db);
   }
   else
